Fix running KFAC with bfloat16 weights (#239)
* ekfac_tests: Read cached index_config.yaml as YAML, not JSON

This code path wasn't updated after the switch to YAML.

* Demonstrate issue with KFAC with bf16 precision

Currently, KFAC does not work well when run with --precision bf16, but the
existing tests are not well-suited to understanding and fixing the issue
properly. This commit therefox fixes two issues:

- we change the ground truth computation of eigenvectors and eigenvalues
to use float32 vectors and operations even if we're running the model itself in
bf16, this way it stays close to the actual ground truth we care about. This
avoids mistakenly thinking the real computation is good enough when it isn't.

- We replace the fragile test_eigenvectors and test_eigenvalue_correction by a
new test_ekfac_recomposition. This is more robust because the eigendecomposition
itself can be very different due to small numerical differences even if the
recomposed matrices are very close. This avoids mistakenly thinking the real
computation is not good enough when it is.

After these two changes, we observe that bf16 support is indeed broken:

```
$ pytest tests/ekfac_tests/test_ekfac_recomposition.py -sv --precision bf16 --model_name EleutherAI/Pythia-14m

FAILED tests/ekfac_tests/test_ekfac_recomposition.py::test_ekfac_recomposition -
AssertionError: EKFAC GNH recomposition: max per-layer rel_error=2.66e-01,
expected < 0.001
```

* Fix running KFAC with bfloat16 weights

This fixes two issues when running a model with --precision bf16:

1. Computing `a.mT @ a` and `g.mT @ g` in bf16 is too inaccurate and leads to
   broken eigenvectors.

2. `torch.einsum` in ShardedMul raised
   `RuntimeError: expected scalar type Float but found BFloat16`
   when the stored eigenvectors (fp32) were combined with the model's bf16
   activations/gradients.

Also get rid of --hessian_dtype "auto" and default to "fp32" since using less
precision doesn't seem to work well in practice.

The test added in the previous commit now passes with the same parameters:

```
$ pytest tests/ekfac_tests/test_ekfac_recomposition.py -sv --precision bf16
--model_name EleutherAI/Pythia-14m
```

Note that with Pythia-14m specifically, the test fails with world_size 8
regardless of the precision:

```
$ pytest tests/ekfac_tests/test_ekfac_recomposition.py -sv
--model_name EleutherAI/Pythia-14m --world_size 8

FAILED tests/ekfac_tests/test_ekfac_recomposition.py::test_ekfac_recomposition -
AssertionError: EKFAC GNH recomposition: max per-layer rel_error=1.98e-01, expected < 0.05

```
$ pytest tests/ekfac_tests/test_ekfac_recomposition.py -sv --precision bf16
--model_name EleutherAI/Pythia-14m --world_size 8
FAILED tests/ekfac_tests/test_ekfac_recomposition.py::test_ekfac_recomposition -
AssertionError: EKFAC GNH recomposition: max per-layer rel_error=6.98e-01, expected < 0.05
```

But it works with allenai/OLMo-2-0425-1B with both float32 and bf16.

* slightly bump up tolerance for test to pass

---------

+++ bergson/hessians/kfac.py
@@ -48,7 +48,7 @@ class CovarianceCollector(HookCollectorBase):
         assert mask is not None, "Valid mask not set for forward hook."
 
         # a: [N, S, I], valid_masks: [N, S] -> select valid positions
-        a_bi = a[mask]  # [num_valid, I]
+        a_bi = a[mask].to(self.dtype)  # [num_valid, I]
 
         # Compute local covariance
         local_update_ii = a_bi.mT @ a_bi
@@ -72,7 +72,7 @@ class CovarianceCollector(HookCollectorBase):
         mask = self._current_valid_mask
 
         # g: [N, S, O], mask: [N, S] -> select valid positions
-        g_bo = g[mask]  # [num_valid, O]
+        g_bo = g[mask].to(self.dtype)  # [num_valid, O]
 
         # Compute local covariance
         local_update_oo = g_bo.mT @ g_bo
